# bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "WK API",
    "description": "We Kare Api 1.0 ",
    "termsOfService": "http://www.wkya.rw",
    "contact": {
      "name": "wecare",
      "url": "http://www.wkya.rw",
      "email": "rwahamanick@gmail.com"
    },
    "version": "v1",
    "x-logo": {
      "url": "/img/tiger-soft-logo.png",
      "backgroundColor": "#FFFFFF",
      "altText": "PetStore Logo"
    }
  },
  "paths": {
...
    "/barbers/partner/{partnerId}": {
      "get": {
        "tags": [
          "Barber"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "partnerId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/partners/{id}": {
      "get": {
        "tags": [
          "Partner"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Partner"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PartnerInputDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Partner"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/partnerServices": {
      "get": {
        "tags": [
          "PartnerService"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "PartnerService"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PartnerServiceInputDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/partnerServices/partner/{partnerId}": {
      "get": {
        "tags": [
          "PartnerService"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "partnerId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/partnerServices/service/{serviceId}": {
      "get": {
        "tags": [
          "PartnerService"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "serviceId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/partnerServices/{id}": {
      "get": {
        "tags": [
          "PartnerService"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "PartnerService"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PartnerServiceInputDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "PartnerService"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/partnerServices/uploadServiceImage": {
      "put": {
        "tags": [
          "PartnerService"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "PartnerServiceId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "ImageData",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/reservations/partner/{partnerId}": {
      "get": {
        "tags": [
          "Reservation"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "partnerId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/roles": {
      "get": {
        "tags": [
          "Role"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Role"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ReferenceInputDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/roles/{id}": {
      "get": {
        "tags": [
          "Role"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Role"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ReferenceInputDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Role"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ReferenceInputDto": {
      "required": [
        "name"
      ],
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
    "PartnerInputDto": {
      "type": "object",
      "properties": {
        "name": {
          "maxLength": 250,
          "type": "string"
        },
        "location": {
          "maxLength": 300,
          "type": "string"
        },
        "email": {
          "maxLength": 125,
          "type": "string"
        },
        "phone": {
          "maxLength": 10,
          "type": "string"
        },
        "momoCode": {
          "maxLength": 8,
          "type": "string"
        },
        "userId": {
          "format": "uuid",
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
...
    "PartnerServiceInputDto": {
      "required": [
        "businessServiceId",
        "partnerId",
        "price"
      ],
      "type": "object",
      "properties": {
        "partnerId": {
          "format": "int32",
          "type": "integer"
        },
        "businessServiceId": {
          "format": "int32",
          "type": "integer"
        },
        "price": {
          "format": "int32",
          "type": "integer"
        },
        "seats": {
          "format": "int32",
          "type": "integer"
        },
        "minDuration": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Authorization"
    }
  },
  "security": [
    { }
  ]
}